Getting prepared to compile for other platforms.
By getting stricter with compilation warnings, some porting bugs
can be early detected (e.g. typical bug is to use an integer
that gets truncated in win32 but not in linux64. It can be difficult
to spot once the code is written.  -Wconversion avoids that).

I had to fix very few lines of boost-process to stick to the new
stricter rules.

+++ client/src/main/c++/build_log.cpp
@@ -22,9 +22,6 @@
 #include "build_log.hpp"
 #include "utils.hpp"
 /////////////////////////////////////////////////////////////////////////
-// BOOST
-#include "boost/process.hpp" 
-/////////////////////////////////////////////////////////////////////////
 // STD 
 #include <iomanip>
 #include <unistd.h>
@@ -180,7 +177,7 @@ magrit::log::colorize_linux ( const std::string& status, bool color )
 {
   std::stringstream output;
 
-  for ( uint i = 0 ; i < status.size() ; ++i )
+  for ( size_t i = 0 ; i < status.size() ; ++i )
   {
     switch ( status[i] )
     {
